+++ engine/flink/components/table/src/main/scala/pl/touk/nussknacker/engine/flink/table/source/TableSourceFactory.scala
@@ -11,6 +11,7 @@ import pl.touk.nussknacker.engine.api.definition._
 import pl.touk.nussknacker.engine.api.parameter.ParameterName
 import pl.touk.nussknacker.engine.api.process.{BasicContextInitializer, Source, SourceFactory}
 import pl.touk.nussknacker.engine.api.{NodeId, Params}
+import pl.touk.nussknacker.engine.flink.table.TableComponentProviderConfig.TestDataGenerationMode
 import pl.touk.nussknacker.engine.flink.table.source.TableSourceFactory.tableNameParamName
 import pl.touk.nussknacker.engine.flink.table.utils.TableComponentFactory
 import pl.touk.nussknacker.engine.flink.table.utils.TableComponentFactory._
@@ -19,7 +20,7 @@ import pl.touk.nussknacker.engine.flink.table.{TableDefinition, TableSqlDefiniti
 class TableSourceFactory(
     definition: TableSqlDefinitions,
     enableFlinkBatchExecutionMode: Boolean,
-    useRealDataForTests: Boolean
+    testDataGenerationMode: TestDataGenerationMode
 ) extends SingleInputDynamicComponent[Source]
     with SourceFactory
     with BoundedStreamComponent {
@@ -52,7 +53,7 @@ class TableSourceFactory(
     val selectedTable = finalStateOpt.getOrElse(
       throw new IllegalStateException("Unexpected (not defined) final state determined during parameters validation")
     )
-    new TableSource(selectedTable, definition.sqlStatements, enableFlinkBatchExecutionMode, useRealDataForTests)
+    new TableSource(selectedTable, definition.sqlStatements, enableFlinkBatchExecutionMode, testDataGenerationMode)
   }
 
   override def nodeDependencies: List[NodeDependency] = List.empty